starfive: enable BOOTCOMMAND so that DISTRO_DEFAULTS becomes useful
Currently, the starfive configuration implements the distro spec:
https://u-boot.readthedocs.io/en/latest/develop/distro.html

However, BOOTCOMMAND is not enabled in the actual board config.  This
results in "bootcmd" not being defined, so u-boot does nothing by default.

The Fedora image provided by StarFive is still able to boot, but only
because u-boot has a preboot command that loads boot/uEnv.txt from the
third partition of the SD card.

To make use of the distro spec and allow generic booting of any OS image,
this patch enables BOOTCOMMAND.

Signed-off-by: Baptiste Jonglez <git@bitsofnetworks.org>