configs: starfive: Change boot order to try DHCP last
Currently, the boot order is DHCP, then MMC.

Trying DHCP boot first is not user-friendly, because it can take a long
time before falling back to MMC.  Also, network booting is usually not
wanted when a SD card is inserted.

Change order so that we first try MMC, and then fall back to DHCP if MMC
booting didn't work.

Signed-off-by: Baptiste Jonglez <git@bitsofnetworks.org>