common: autoboot: Add auto-fastboot on StarFive SoC
Add auto-fastboot function before autoboot on StarFive SoC
and add a config to enable or disable this.

In the auto-fastboot, it check a environment flag of 'fb_sf_completed'
first. The flag is NULL and then the fastboot will work.
The flag will be set from USB tool to indicate done after flashing the
image.

Signed-off-by: Xingyu Wu <xingyu.wu@starfivetech.com>