include: asm: arch-jhb100: [WORKAROUND] Add sys and bmcperiph0 C macro
Add the sys and bmcperiph0 C macro to support pin control (gate/ungate),
reset (assert/deassert), and clock (enable/disable) functionality. These
additions are essential for ensuring proper operation of these features
while the drivers are still under development. This is particularly
important for the UART driver, which is needed to communicate and execute
commands during the Linux bring-up process.

Signed-off-by: zhiwei.lim <zhiwei.lim@starfivetech.com>