include: asm: arch-jhb100: Disable SMBus filter using C macro
The SMBus filter is enabled by default, which blocks I2C write
operations. To temporarily address this issue, a C macro is used
to disable the filter.

This is a temporary workaround until the filter functionality is
fully implemented in the L1 firmware. Once the L1 firmware is ready
and the filter implementation is complete, this macro will be removed.

Signed-off-by: zhiwei.lim <zhiwei.lim@starfivetech.com>