drivers: net: dwc_eth_qos_starfive.c: Remove reset assert
Changes:
1. Change API callback for SGMII innosilicon PHY for proper shutdown.
2. Remove GMAC controller reset assert in the stop-reset callback.

Asserting the JHB100 GMAC controller is unsafe, as some
framework paths may still access its registers, leading to
a CPU hang.

For example, TFTP fails to find the file:
/net/tftp.c: line 681: eth_halt()

Calling eth_halt() eventually triggers the stop-reset callback.

In /drivers/net/dwc_eth_qos.c: line 1180,
a write to a GMAC DMA register causes a CPU hang when the
controller is in the asserted reset state.

This occurs because the GMAC controller does not
respond while in reset. The CPU then waits indefinitely
for a response, resulting in a system hang.

Signed-off-by: John Chew <yuinyee.chew@starfivetech.com>